Re-word project focus and intent
It does not really make sense to rebuilt the whole web ui.
Instead, focus on dealing with issues first.

This is a mundane task for developers, that could be supported
with the help of a cli.

It should be satisfying to (finally?) be able to close an issue.
I imagine doing stuff like 'glab -c 117' to close it or
'glab -rm 117' to delete it and "'glab -rm 'Doing' 'In Review'" to
clean up labels.