Add new features
- print markdown in terminal
- align labels on issues
- improve scaffolding

This project was started on a relatively boring evening,
when I just wanted to code something. I took out my list
with potential projects and this was the one that I felt
I could get the most personal benefit from short-term.

Later, I realized this might also be a project worth
introducing others to, e.g. during a barcamp. Therefore,
I am making these changes now without properly splitting
the commits.

Largely because the development was sort of a trial
and error approach, in order to become well prepared for
potential questions that might arise. So there really were
no commits that surfaced just by working on `glab`, because
I changed things all over to break it, learn from it and
make my session an awesome one.

Unfortunately, the state of the project was not good but
I wanted to host a session anyway. So there it is, this is
why this commit